+++ src/main/java/com/example/smartorder/service/admin/Impl/AdminMemberServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.smartorder.service.admin;
+package com.example.smartorder.service.admin.Impl;
 
 import static com.example.smartorder.common.error.ErrorCode.NOT_FOUND_USER;
