+++ src/main/java/com/example/smartorder/service/admin/Impl/AdminStoreMenuServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.smartorder.service.admin;
+package com.example.smartorder.service.admin.Impl;
 
 import static com.example.smartorder.common.error.ErrorCode.NOT_FOUND_STORE;
 import static com.example.smartorder.type.SaleState.ON_SALE;
